requirements.txt
# ===============================
# Core ML & Data Processing
# ===============================
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
scipy==1.12.0
xgboost==2.0.3
lightgbm==4.3.0

# ===============================
# Neural Networks
# ===============================
tensorflow==2.15.0
torch==2.1.2

# ===============================
# Explainability
# ===============================
shap==0.44.1
lime==0.2.0.1
plotly==5.19.0

# ===============================
# API & Web Framework
# ===============================
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
waitress==2.1.2

# ===============================
# Data Quality & Drift
# ===============================
evidently==0.4.24
distance==0.1.3

# ===============================
# Monitoring & Logging
# ===============================
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1

# ===============================
# Utilities
# ===============================
joblib==1.3.2
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
requests>=2.32.2,<3.0.0
python-dotenv==1.0.1

# ===============================
# Development
# ===============================
pytest==7.4.4
black==24.1.1
flake8==6.1.0
